MAILBOX-273 move MailboxId to mailbox-api project
