+++ mailbox/api/src/main/java/org/apache/james/mailbox/model/MailboxId.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
-package org.apache.james.mailbox.store.mail.model;
+package org.apache.james.mailbox.model;
 
 public interface MailboxId {
     String serialize();
