+++ mailbox/maildir/src/main/java/org/apache/james/mailbox/maildir/MaildirId.java
@@ -18,7 +18,7 @@
  ****************************************************************/
 package org.apache.james.mailbox.maildir;
 
-import org.apache.james.mailbox.store.mail.model.MailboxId;
+import org.apache.james.mailbox.model.MailboxId;
 
 public class MaildirId implements MailboxId {
